Relocate Logger type
