+++ src/Connection.ts
@@ -40,11 +40,19 @@ import type {
 	DbSubroutineResponseTypes,
 	DbSubroutineResponseTypesMap,
 	GenericObject,
-	Logger,
 } from './types';
 import { dummyLogger } from './utils';
 
 // #region Types
+export interface Logger {
+	error(message: string): void;
+	warn(message: string): void;
+	info(message: string): void;
+	verbose(message: string): void;
+	debug(message: string): void;
+	silly(message: string): void;
+}
+
 export interface CreateConnectionOptions {
 	/** Optional logger instance */
 	logger?: Logger;
