+++ src/types/index.ts
@@ -28,12 +28,3 @@ export type EncryptFunc = {
 	(data: unknown | null): string | null;
 	(data: (unknown | null)[]): (string | null)[];
 };
-
-export interface Logger {
-	error(message: string): void;
-	warn(message: string): void;
-	info(message: string): void;
-	verbose(message: string): void;
-	debug(message: string): void;
-	silly(message: string): void;
-}
