+++ src/utils/__tests__/dummyLogger.test.ts
@@ -1,4 +1,4 @@
-import type { Logger } from '../../types';
+import type { Logger } from '../../Connection';
 import dummyLogger from '../dummyLogger';
 
 test('should return void for all expected log levels', () => {
